backend/flask_api/requirements.txt
flask==2.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==9.5.0
python-dotenv==0.19.0
Werkzeug==2.0.1 